Simplified APIs. Incorporated PR comments.
