+++ libs/server/Module/ModuleRegistrar.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
@@ -12,16 +11,16 @@ using Microsoft.Extensions.Logging;
 namespace Garnet.server.Module
 {
     /// <summary>
-    /// Common interface that all Garnet modules must implement.
+    /// Abstract base class that all Garnet modules must inherit from.
     /// </summary>
-    public interface IModule
+    public abstract class ModuleBase
     {
         /// <summary>
         /// Called when the module is loaded.
         /// </summary>
         /// <param name="context">Module load context</param>
-        /// <param name="moduleArgs">Module load args</param>
-        void OnLoad(ModuleLoadContext context, List<string> moduleArgs);
+        /// <param name="args">Module load args</param>
+        public abstract void OnLoad(ModuleLoadContext context, string[] args);
     }
 
     /// <summary>
@@ -41,14 +40,15 @@ namespace Garnet.server.Module
     /// </summary>
     public class ModuleLoadContext
     {
+        public readonly ILogger Logger;
         internal string Name;
         internal uint Version;
         internal bool Initialized;
 
-        private ModuleRegistrar moduleRegistrar;
-        private CustomCommandManager customCommandManager;
+        private readonly ModuleRegistrar moduleRegistrar;
+        private readonly CustomCommandManager customCommandManager;
 
-        internal ModuleLoadContext(ModuleRegistrar moduleRegistrar, CustomCommandManager customCommandManager)
+        internal ModuleLoadContext(ModuleRegistrar moduleRegistrar, CustomCommandManager customCommandManager, ILogger logger)
         {
             Debug.Assert(moduleRegistrar != null);
             Debug.Assert(customCommandManager != null);
@@ -56,6 +56,7 @@ namespace Garnet.server.Module
             Initialized = false;
             this.moduleRegistrar = moduleRegistrar;
             this.customCommandManager = customCommandManager;
+            Logger = logger;
         }
 
         /// <summary>
@@ -92,13 +93,9 @@ namespace Garnet.server.Module
         /// <param name="commandInfo">Command info</param>
         /// <param name="expirationTicks">Expiration ticks for the key</param>
         /// <returns>Registration status</returns>
-        public ModuleActionStatus RegisterCommand(string name, int numParams, CommandType type, CustomRawStringFunctions customFunctions, RespCommandsInfo commandInfo, long expirationTicks = 0)
+        public ModuleActionStatus RegisterCommand(string name, CustomRawStringFunctions customFunctions, CommandType type = CommandType.ReadModifyWrite, int numParams = int.MaxValue, RespCommandsInfo commandInfo = null, long expirationTicks = 0)
         {
-            if (string.IsNullOrEmpty(name))
-                return ModuleActionStatus.InvalidRegistrationInfo;
-            if (customFunctions == null)
-                return ModuleActionStatus.InvalidRegistrationInfo;
-            if (commandInfo == null)
+            if (string.IsNullOrEmpty(name) || customFunctions == null)
                 return ModuleActionStatus.InvalidRegistrationInfo;
 
             customCommandManager.Register(name, numParams, type, customFunctions, commandInfo, expirationTicks);
@@ -114,11 +111,9 @@ namespace Garnet.server.Module
         /// <param name="proc">Transaction procedure implemenation</param>
         /// <param name="commandInfo">Command info</param>
         /// <returns>Registration status</returns>
-        public ModuleActionStatus RegisterTransaction(string name, int numParams, Func<CustomTransactionProcedure> proc, RespCommandsInfo commandInfo = null)
+        public ModuleActionStatus RegisterTransaction(string name, Func<CustomTransactionProcedure> proc, int numParams = int.MaxValue, RespCommandsInfo commandInfo = null)
         {
-            if (string.IsNullOrEmpty(name))
-                return ModuleActionStatus.InvalidRegistrationInfo;
-            if (proc == null)
+            if (string.IsNullOrEmpty(name) || proc == null)
                 return ModuleActionStatus.InvalidRegistrationInfo;
 
             customCommandManager.Register(name, numParams, proc, commandInfo);
@@ -140,14 +135,14 @@ namespace Garnet.server.Module
 
         private readonly ConcurrentDictionary<string, ModuleLoadContext> modules;
 
-        internal bool LoadModule(CustomCommandManager customCommandManager, Assembly loadedAssembly, List<string> moduleArgs, ILogger logger, out ReadOnlySpan<byte> errorMessage)
+        internal bool LoadModule(CustomCommandManager customCommandManager, Assembly loadedAssembly, string[] moduleArgs, ILogger logger, out ReadOnlySpan<byte> errorMessage)
         {
             errorMessage = default;
 
             // Get types that implement IModule from loaded assemblies
             var loadedTypes = loadedAssembly
                 .GetTypes()
-                .Where(t => t.GetInterfaces().Contains(typeof(IModule)))
+                .Where(t => t.IsSubclassOf(typeof(ModuleBase)) && !t.IsAbstract)
                 .ToArray();
 
             if (loadedTypes.Length == 0)
@@ -170,26 +165,17 @@ namespace Garnet.server.Module
                 return false;
             }
 
-            var instance = (IModule)Activator.CreateInstance(moduleType);
+            var instance = (ModuleBase)Activator.CreateInstance(moduleType);
             if (instance == null)
             {
                 errorMessage = CmdStrings.RESP_ERR_GENERIC_INSTANTIATING_CLASS;
                 return false;
             }
 
+            var moduleLoadContext = new ModuleLoadContext(this, customCommandManager, logger);
             try
             {
-                var moduleLoadContext = new ModuleLoadContext(this, customCommandManager);
                 instance.OnLoad(moduleLoadContext, moduleArgs);
-
-                if (!moduleLoadContext.Initialized)
-                {
-                    errorMessage = CmdStrings.RESP_ERR_MODULE_ONLOAD;
-                    return false;
-                }
-
-                logger?.LogInformation("Module {0} version {1} loaded", moduleLoadContext.Name, moduleLoadContext.Version);
-                return true;
             }
             catch (Exception ex)
             {
@@ -197,6 +183,16 @@ namespace Garnet.server.Module
                 errorMessage = CmdStrings.RESP_ERR_MODULE_ONLOAD;
                 return false;
             }
+
+            if (!moduleLoadContext.Initialized)
+            {
+                errorMessage = CmdStrings.RESP_ERR_MODULE_ONLOAD;
+                logger?.LogError("Module {0} failed to initialize", moduleLoadContext.Name);
+                return false;
+            }
+
+            logger?.LogInformation("Module {0} version {1} loaded", moduleLoadContext.Name, moduleLoadContext.Version);
+            return true;
         }
 
         internal bool TryAdd(ModuleLoadContext moduleLoadContext)
