+++ playground/CommandInfoUpdater/GarnetCommandsInfo.json
@@ -482,5 +482,34 @@
     "Tips": null,
     "KeySpecifications": null,
     "SubCommands": null
+  },
+  {
+    "Command": "MODULE",
+    "Name": "MODULE",
+    "IsInternal": false,
+    "Arity": -2,
+    "Flags": "Admin",
+    "FirstKey": 0,
+    "LastKey": 0,
+    "Step": 0,
+    "AclCategories": "Admin, Dangerous, Slow",
+    "Tips": null,
+    "KeySpecifications": null,
+    "SubCommands": [
+      {
+        "Command": "MODULE",
+        "Name": "MODULE|LOADCS",
+        "IsInternal": false,
+        "Arity": -3,
+        "Flags": "Admin, NoAsyncLoading, NoScript",
+        "FirstKey": 0,
+        "LastKey": 0,
+        "Step": 0,
+        "AclCategories": "Admin, Dangerous, Slow",
+        "Tips": null,
+        "KeySpecifications": null,
+        "SubCommands": null
+      }
+    ]
   }
 ]
