+++ test/Garnet.test/RespModuleTests.cs
@@ -60,21 +60,22 @@ namespace Garnet.test
             using (var testFile = File.CreateText(testFilePath))
             {
                 testFile.WriteLine(
-                    "using Garnet.server.Module;" +
-                    "using System.Collections.Generic;" +
-                    "using Tsavorite.core;" +
-                    "using Garnet;" +
-                    "using Garnet.server;" +
-                    "namespace TestGarnetModule " +
-                    "{ " +
-                    "   public class TestModule : IModule " +
-                    "   { " +
-                    "       public void OnLoad(ModuleLoadContext context, List<string> moduleArgs)" +
-                    "       { " +
+                    @"using System;
+                    using System.Collections.Generic; 
+                    using Garnet; 
+                    using Garnet.server; 
+                    using Garnet.server.Module; 
+                    using Tsavorite.core; 
+                    namespace TestGarnetModule  
+                    {  
+                       public class TestModule : ModuleBase  
+                       {  
+                           public override void OnLoad(ModuleLoadContext context, string[] args) 
+                           {" +
                     onLoadBody +
-                    "       } " +
-                    "   } " +
-                    "}");
+                    @"     }
+                       }
+                    }");
             }
 
             var modulePath = Path.Combine(dir1, "TestModule.dll");
@@ -90,20 +91,22 @@ namespace Garnet.test
         public void TestModuleLoad()
         {
             var onLoad =
-                    "            context.Initialize(\"TestModule\", 1); " +
-                    "            context.RegisterCommand(\"TestModule.SetIfPM\", 2, CommandType.ReadModifyWrite, new SetIfPMCustomCommand()," +
-                    "            new RespCommandsInfo{ Name = \"TestModule.SETIFPM\", Arity = 4, FirstKey = 1, LastKey = 1, Step = 1," +
-                    "            Flags = RespCommandFlags.DenyOom | RespCommandFlags.Write, AclCategories = RespAclCategories.String | RespAclCategories.Write}); " +
-                    "            context.RegisterTransaction(\"TestModule.READWRITETX\", 3, () => new ReadWriteTxn()," +
-                    "            new RespCommandsInfo{ Name = \"TestModule.READWRITETX\", Arity = 4, FirstKey = 1, LastKey = 3, Step = 1," +
-                    "            Flags = RespCommandFlags.DenyOom | RespCommandFlags.Write, AclCategories = RespAclCategories.Write}); ";
+                    @"context.Initialize(""TestModule"", 1);
+                    
+                    context.RegisterCommand(""TestModule.SetIfPM"", new SetIfPMCustomCommand(), CommandType.ReadModifyWrite, 2,
+                    new RespCommandsInfo { Name = ""TestModule.SETIFPM"", Arity = 4, FirstKey = 1, LastKey = 1, Step = 1,
+                    Flags = RespCommandFlags.DenyOom | RespCommandFlags.Write, AclCategories = RespAclCategories.String | RespAclCategories.Write });
+                    
+                    context.RegisterTransaction(""TestModule.READWRITETX"", () => new ReadWriteTxn(), 3,
+                    new RespCommandsInfo { Name = ""TestModule.READWRITETX"", Arity = 4, FirstKey = 1, LastKey = 3, Step = 1,
+                    Flags = RespCommandFlags.DenyOom | RespCommandFlags.Write, AclCategories = RespAclCategories.Write });";
 
             var modulePath = CreateTestModule(onLoad);
 
             using var redis = ConnectionMultiplexer.Connect(TestUtils.GetConfig());
             var db = redis.GetDatabase(0);
 
-            var resp = db.Execute($"MODULE", ["LOAD", modulePath]);
+            var resp = db.Execute($"MODULE", ["LOADCS", modulePath]);
             Assert.AreEqual("OK", (string)resp);
 
             // Test SETIFPM
@@ -136,6 +139,30 @@ namespace Garnet.test
         }
 
         [Test]
+        public void TestModuleLoadCSArgs()
+        {
+            var onLoad =
+                    @"context.Initialize(""TestModuleLoadCSArgs"", 1);
+                    
+                    if (args.Length != 2)
+                        throw new Exception(""Invalid number of arguments"");
+
+                    if (args[0] != ""arg0"")
+                        throw new Exception($""Incorrect arg value {args[0]}"");
+
+                    if (args[1] != ""arg1"")
+                        throw new Exception($""Incorrect arg value {args[1]}"");";
+
+            var modulePath = CreateTestModule(onLoad);
+
+            using var redis = ConnectionMultiplexer.Connect(TestUtils.GetConfig());
+            var db = redis.GetDatabase(0);
+
+            var resp = db.Execute($"MODULE", ["LOADCS", modulePath, "arg0", "arg1"]);
+            Assert.AreEqual("OK", (string)resp);
+        }
+
+        [Test]
         public void TestUninitializedModule()
         {
             var modulePath = CreateTestModule("");
@@ -145,8 +172,29 @@ namespace Garnet.test
 
             try
             {
-                db.Execute($"MODULE", ["LOAD", modulePath]);
+                db.Execute($"MODULE", ["LOADCS", modulePath]);
                 Assert.Fail("Module with empty OnLoad should not load successfully");
+            }
+            catch (RedisException ex)
+            {
+                Assert.AreEqual("ERR Error during module OnLoad", ex.Message);
+            }
+        }
+
+        [Test]
+        public void TestAlreadyLoadedModule()
+        {
+            var modulePath = CreateTestModule(
+                @"context.Initialize(""TestAlreadyLoadedModule"", 1);");
+
+            using var redis = ConnectionMultiplexer.Connect(TestUtils.GetConfig());
+            var db = redis.GetDatabase(0);
+
+            db.Execute($"MODULE", ["LOADCS", modulePath]);
+            try
+            {
+                db.Execute($"MODULE", ["LOADCS", modulePath]);
+                Assert.Fail("Already loaded module should not successfully load again");
             }
             catch (RedisException ex)
             {
@@ -165,7 +213,7 @@ namespace Garnet.test
 
             try
             {
-                db.Execute($"MODULE", ["LOAD", nonExistentModulePath]);
+                db.Execute($"MODULE", ["LOADCS", nonExistentModulePath]);
                 Assert.Fail("Non existing module assembly should not load successfully");
             }
             catch (RedisException ex)
